fix(tarko): simplify shared header component to reduce code duplication
